Add Okta UsernameSuffix configuration option to dis-ambiguate users.
This solves the problem when an Okta domain has multiple "subdomains" with
the same username in different subdomains referring to different identities
and hence different passwords.

For example, "fred@domain.com" is unable to log in because
"fred@altdomain.com" was created first and Okta is expecting the password
for "fred@altdomain" when "fred" (@domain.com) attempts to log in.

This collision can happen because:

- A single Okta domain has multiple with "subdomains"

- Keymaster is stripping the `@` and everything afterwards (for internal
  consistency)