Hide legacy Authenticate and Register buttons.
Hide the legacy Authenticate button. This is no longer needed because all
modern browsers support FIDOv2 and the compatibility code supports
devices registered with FIDOv1.

Hide the legacy Register token button if password login is not enabled.
Sites using only federated login do not need to provide compatibility
for the CLI (which only supports FIDOv1 at the moment), as they redirect
to the Web browser for logins.