Role requesting certs (#251)
This change adds 2 endpoints to keymasterd. A role requesting ednpoint and a refresh endpoint. 

The first of this endpoints (role requesting) requires an "automationadmin" (new option in the config) to be the caller. This endpoint generates much longer lived certs (today hardcoded to 45 days) and thus we have also enhanced the verification of identities to include denylists of keys. These are currenly part of the keymasterd configuration 
Tthe second endpoint can only be called from the ipranges embedded in the certificate and provide a mechanism to upate things.